requirements.txt
python-dotenv 
langchain 
langchain-community 
langchain-openai 
langchainhub 
openai 
tiktoken 
azure-ai-documentintelligence 
azure-identity 
azure-search-documents==11.6.0b3